+++ packages/data-schema/src/runtime/client/index.ts
@@ -18,6 +18,9 @@ import type {
   StringFilter,
   NumericFilter,
   BooleanFilters,
+  SubscriptionStringFilter,
+  SubscriptionNumericFilter,
+  SubscriptionBooleanFilters,
 } from '../../util';
 import { AmplifyServer } from '../bridge-types';
 
@@ -419,6 +422,26 @@ type ModelFilter<
       : StringFilter;
 };
 
+type LogicalSubscriptionFilters<
+  Model extends ClientSchemaByEntityTypeBaseShape['models'][string],
+> = {
+  and?: ModelSubscriptionFilter<Model> | ModelSubscriptionFilter<Model>[];
+  or?: ModelSubscriptionFilter<Model> | ModelSubscriptionFilter<Model>[];
+  not?: ModelSubscriptionFilter<Model>;
+};
+
+type ModelSubscriptionFilter<
+  Model extends ClientSchemaByEntityTypeBaseShape['models'][string],
+> = LogicalSubscriptionFilters<Model> & {
+  [K in keyof Model['type'] as Model['type'][K] extends LazyLoader<any, any>
+    ? never
+    : K]?: boolean extends Model['type'][K]
+    ? SubscriptionBooleanFilters
+    : number extends Model['type'][K]
+      ? SubscriptionNumericFilter
+      : SubscriptionStringFilter;
+};
+
 export type ModelSortDirection = 'ASC' | 'DESC';
 
 type ListCpkOptions<
@@ -532,7 +555,7 @@ type ModelTypesClient<
   onCreate<
     SelectionSet extends ReadonlyArray<ModelPath<FlatModel>> = never[],
   >(options?: {
-    filter?: ModelFilter<Model>;
+    filter?: ModelSubscriptionFilter<Model>;
     selectionSet?: SelectionSet;
     authMode?: AuthMode;
     authToken?: string;
@@ -543,7 +566,7 @@ type ModelTypesClient<
   onUpdate<
     SelectionSet extends ReadonlyArray<ModelPath<FlatModel>> = never[],
   >(options?: {
-    filter?: ModelFilter<Model>;
+    filter?: ModelSubscriptionFilter<Model>;
     selectionSet?: SelectionSet;
     authMode?: AuthMode;
     authToken?: string;
@@ -554,7 +577,7 @@ type ModelTypesClient<
   onDelete<
     SelectionSet extends ReadonlyArray<ModelPath<FlatModel>> = never[],
   >(options?: {
-    filter?: ModelFilter<Model>;
+    filter?: ModelSubscriptionFilter<Model>;
     selectionSet?: SelectionSet;
     authMode?: AuthMode;
     authToken?: string;